Add CI pipeline to build patcher
Modifies the makefile to play nicely with github actions. Notably by surrounding the pushes to env.ld in quotes in the makefile so they are not interpreted as a separate line.

Also removes the infinite loop from the patcher to make it more CI friendly.

MinGW at the time of the patch comes preinstalled at version 8.1.0 on windows-latest. A version check has been added to force the MinGW version as desired by using choco windows package manager.